JWT token implementation and refresh token

+++ resources/item.py
@@ -4,23 +4,30 @@ from models import ItemModel
 from resources.schemas import ItemSchema, ItemUpdateSchema
 from db import db
 from sqlalchemy.exc import SQLAlchemyError
+from flask_jwt_extended import jwt_required, get_jwt
 
 blp = Blueprint("Items", __name__, description="Item API")
 
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
+    @blp.alt_response(404, description="Item not found.")
     def delete(self, item_id):
-        item = ItemModel.query.get_or_404(item_id)
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin previlege is required")
         try:
+            item = ItemModel.query.get_or_404(item_id)
             db.session.delete(item)
             db.session.commit()
-            return {"message": "Item deleted."}
+            return {"message": "Item deleted."}, 200
         except SQLAlchemyError:
             abort(500, message="An error occurred while deleting the item.")
 
@@ -61,6 +68,7 @@ class ItemView(MethodView):
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
